+++ src/main/java/blusunrize/immersiveengineering/common/register/IEBannerPatterns.java
@@ -32,6 +32,7 @@ public class IEBannerPatterns
 	public static final List<BannerEntry> ALL_BANNERS = new ArrayList<>();
 	public static final BannerEntry HAMMER = addBanner("hammer", "hmr", "head","grip");
 	public static final BannerEntry WIRECUTTER = addBanner("wirecutter", "wct", "head","grip");
+	public static final BannerEntry SCREWDRIVER = addBanner("screwdriver", "scd", "head","grip");
 	public static final BannerEntry BEVELS = addBanner("bevels", "bvl");
 	public static final BannerEntry ORNATE = addBanner("ornate", "orn");
 	public static final BannerEntry TREATED_WOOD = addBanner("treated_wood", "twd");