+++ changelog.md
@@ -4,6 +4,7 @@
     - There's also matching trapdoors
 - Add the redstone-controlled pipe valve from Engineer's Decor (voidsong-dragonfly)
     - Thank you wilechaote for giving us permission to include these blocks in the mod <3
+- Add a banner pattern for the screwdriver (BluSunrize)
 - Add a recipe for armor-piercing cartridges using netherite nuggets (BluSunrize)
 - Add the ability to open scaffold covered ladders on the side (voidsong-dragonfly)
 - Add lots more block to be dug by the grinding disk, rock cutter and engineer's hammer (voidsong-dragonfly)